keep GetIngressHost in cliutils
