+++ pkg/cliutil/uri.go
@@ -1,11 +1,20 @@
 package cliutil
 
 import (
+	"bytes"
+	"fmt"
 	"io"
 	"net/http"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"strings"
+
+	"github.com/solo-io/go-utils/kubeutils"
+	v1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+	"k8s.io/client-go/kubernetes"
 
 	"github.com/solo-io/solo-kit/pkg/errors"
 )
@@ -40,4 +49,107 @@ func GetResource(uri string) (io.ReadCloser, error) {
 
 	// Write the body to file
 	return file, nil
+}
+
+func GetIngressHost(proxyName, proxyNamespace, proxyPort string, localCluster bool) (string, error) {
+	restCfg, err := kubeutils.GetConfig("", "")
+	if err != nil {
+		return "", errors.Wrapf(err, "getting kube rest config")
+	}
+	kube, err := kubernetes.NewForConfig(restCfg)
+	if err != nil {
+		return "", errors.Wrapf(err, "starting kube client")
+	}
+	svc, err := kube.CoreV1().Services(proxyNamespace).Get(proxyName, metav1.GetOptions{})
+	if err != nil {
+		return "", errors.Wrapf(err, "could not detect '%v' service in %v namespace. "+
+			"Check that Gloo has been installed properly and is running with 'kubectl get pod -n gloo-system'",
+			proxyName, proxyNamespace)
+	}
+	var svcPort *v1.ServicePort
+	switch len(svc.Spec.Ports) {
+	case 0:
+		return "", errors.Errorf("service %v is missing ports", proxyName)
+	case 1:
+		svcPort = &svc.Spec.Ports[0]
+	default:
+		for _, p := range svc.Spec.Ports {
+			if p.Name == proxyPort {
+				svcPort = &p
+				break
+			}
+		}
+		if svcPort == nil {
+			return "", errors.Errorf("named port %v not found on service %v", proxyPort, proxyName)
+		}
+	}
+
+	var host, port string
+	// gateway-proxy is an externally load-balanced service
+	if len(svc.Status.LoadBalancer.Ingress) == 0 || localCluster {
+		// assume nodeport on kubernetes
+		// TODO: support more types of NodePort services
+		host, err = getNodeIp(svc, kube)
+		if err != nil {
+			return "", errors.Wrapf(err, "")
+		}
+		port = fmt.Sprintf("%v", svcPort.NodePort)
+	} else {
+		host = svc.Status.LoadBalancer.Ingress[0].Hostname
+		if host == "" {
+			host = svc.Status.LoadBalancer.Ingress[0].IP
+		}
+		port = fmt.Sprintf("%v", svcPort.Port)
+	}
+	return host + ":" + port, nil
+}
+
+func getNodeIp(svc *v1.Service, kube kubernetes.Interface) (string, error) {
+	// pick a node where one of our pods is running
+	pods, err := kube.CoreV1().Pods(svc.Namespace).List(metav1.ListOptions{
+		LabelSelector: labels.SelectorFromSet(svc.Spec.Selector).String(),
+	})
+	if err != nil {
+		return "", err
+	}
+	var nodeName string
+	for _, pod := range pods.Items {
+		if pod.Spec.NodeName != "" {
+			nodeName = pod.Spec.NodeName
+			break
+		}
+	}
+	if nodeName == "" {
+		return "", errors.Errorf("no node found for %v's pods. ensure at least one pod has been deployed "+
+			"for the %v service", svc.Name, svc.Name)
+	}
+	// special case for minikube
+	// we run `minikube ip` which avoids an issue where
+	// we get a NAT network IP when the minikube provider is virtualbox
+	if nodeName == "minikube" {
+		return minikubeIp()
+	}
+
+	node, err := kube.CoreV1().Nodes().Get(nodeName, metav1.GetOptions{})
+	if err != nil {
+		return "", err
+	}
+
+	for _, addr := range node.Status.Addresses {
+		return addr.Address, nil
+	}
+
+	return "", errors.Errorf("no active addresses found for node %v", node.Name)
+}
+
+func minikubeIp() (string, error) {
+	minikubeCmd := exec.Command("minikube", "ip")
+
+	hostname := &bytes.Buffer{}
+
+	minikubeCmd.Stdout = hostname
+	minikubeCmd.Stderr = os.Stderr
+	err := minikubeCmd.Run()
+
+	return strings.TrimSuffix(hostname.String(), "\n"), err
 }
