+++ projects/gloo/cli/pkg/cmd/gateway/url.go
@@ -1,22 +1,13 @@
 package gateway
 
 import (
-	"bytes"
 	"fmt"
-	"os"
-	"os/exec"
-	"strings"
 
-	"github.com/solo-io/go-utils/errors"
-	"github.com/solo-io/go-utils/kubeutils"
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/client-go/kubernetes"
+	"github.com/solo-io/gloo/pkg/cliutil"
 
 	"github.com/solo-io/gloo/projects/gloo/cli/pkg/cmd/options"
 	"github.com/solo-io/go-utils/cliutils"
 	"github.com/spf13/cobra"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func urlCmd(opts *options.Options, optionsFunc ...cliutils.OptionsFunc) *cobra.Command {
@@ -25,7 +16,7 @@ func urlCmd(opts *options.Options, optionsFunc ...cliutils.OptionsFunc) *cobra.C
 		Short: "print the http endpoint for a proxy",
 		Long:  "Use this command to view the HTTP URL of a Proxy reachable from outside the cluster. You can connect to this address from a host on the same network (such as the host machine, in the case of minikube/minishift).",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			ingressHost, err := GetIngressHost(&opts.Proxy, opts.Metadata.Namespace)
+			ingressHost, err := cliutil.GetIngressHost(opts.Proxy.Name, opts.Metadata.Namespace, opts.Proxy.Port, opts.Proxy.LocalCluster)
 			if err != nil {
 				return err
 			}
@@ -51,7 +42,7 @@ func addressCmd(opts *options.Options, optionsFunc ...cliutils.OptionsFunc) *cob
 		Short: "print the socket address for a proxy",
 		Long:  "Use this command to view the address (host:port) of a Proxy reachable from outside the cluster. You can connect to this address from a host on the same network (such as the host machine, in the case of minikube/minishift).",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			ingressHost, err := GetIngressHost(&opts.Proxy, opts.Metadata.Namespace)
+			ingressHost, err := cliutil.GetIngressHost(opts.Proxy.Name, opts.Metadata.Namespace, opts.Proxy.Port, opts.Proxy.LocalCluster)
 			if err != nil {
 				return err
 			}
@@ -65,107 +56,4 @@ func addressCmd(opts *options.Options, optionsFunc ...cliutils.OptionsFunc) *cob
 			"to true if LoadBalanced services are not assigned external IPs by your cluster")
 	cliutils.ApplyOptions(cmd, optionsFunc)
 	return cmd
-}
-
-func GetIngressHost(opts *options.Proxy, namespace string) (string, error) {
-	restCfg, err := kubeutils.GetConfig("", "")
-	if err != nil {
-		return "", errors.Wrapf(err, "getting kube rest config")
-	}
-	kube, err := kubernetes.NewForConfig(restCfg)
-	if err != nil {
-		return "", errors.Wrapf(err, "starting kube client")
-	}
-	svc, err := kube.CoreV1().Services(namespace).Get(opts.Name, metav1.GetOptions{})
-	if err != nil {
-		return "", errors.Wrapf(err, "could not detect '%v' service in %v namespace. "+
-			"Check that Gloo has been installed properly and is running with 'kubectl get pod -n gloo-system'",
-			opts.Name, namespace)
-	}
-	var svcPort *v1.ServicePort
-	switch len(svc.Spec.Ports) {
-	case 0:
-		return "", errors.Errorf("service %v is missing ports", opts.Name)
-	case 1:
-		svcPort = &svc.Spec.Ports[0]
-	default:
-		for _, p := range svc.Spec.Ports {
-			if p.Name == opts.Port {
-				svcPort = &p
-				break
-			}
-		}
-		if svcPort == nil {
-			return "", errors.Errorf("named port %v not found on service %v", opts.Port, opts.Name)
-		}
-	}
-
-	var host, port string
-	// gateway-proxy is an externally load-balanced service
-	if len(svc.Status.LoadBalancer.Ingress) == 0 || opts.LocalCluster {
-		// assume nodeport on kubernetes
-		// TODO: support more types of NodePort services
-		host, err = getNodeIp(svc, kube)
-		if err != nil {
-			return "", errors.Wrapf(err, "")
-		}
-		port = fmt.Sprintf("%v", svcPort.NodePort)
-	} else {
-		host = svc.Status.LoadBalancer.Ingress[0].Hostname
-		if host == "" {
-			host = svc.Status.LoadBalancer.Ingress[0].IP
-		}
-		port = fmt.Sprintf("%v", svcPort.Port)
-	}
-	return host + ":" + port, nil
-}
-
-func getNodeIp(svc *v1.Service, kube kubernetes.Interface) (string, error) {
-	// pick a node where one of our pods is running
-	pods, err := kube.CoreV1().Pods(svc.Namespace).List(metav1.ListOptions{
-		LabelSelector: labels.SelectorFromSet(svc.Spec.Selector).String(),
-	})
-	if err != nil {
-		return "", err
-	}
-	var nodeName string
-	for _, pod := range pods.Items {
-		if pod.Spec.NodeName != "" {
-			nodeName = pod.Spec.NodeName
-			break
-		}
-	}
-	if nodeName == "" {
-		return "", errors.Errorf("no node found for %v's pods. ensure at least one pod has been deployed "+
-			"for the %v service", svc.Name, svc.Name)
-	}
-	// special case for minikube
-	// we run `minikube ip` which avoids an issue where
-	// we get a NAT network IP when the minikube provider is virtualbox
-	if nodeName == "minikube" {
-		return minikubeIp()
-	}
-
-	node, err := kube.CoreV1().Nodes().Get(nodeName, metav1.GetOptions{})
-	if err != nil {
-		return "", err
-	}
-
-	for _, addr := range node.Status.Addresses {
-		return addr.Address, nil
-	}
-
-	return "", errors.Errorf("no active addresses found for node %v", node.Name)
-}
-
-func minikubeIp() (string, error) {
-	minikubeCmd := exec.Command("minikube", "ip")
-
-	hostname := &bytes.Buffer{}
-
-	minikubeCmd.Stdout = hostname
-	minikubeCmd.Stderr = os.Stderr
-	err := minikubeCmd.Run()
-
-	return strings.TrimSuffix(hostname.String(), "\n"), err
 }
